ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sentence-transformers==2.3.1
numpy<2.0.0
mysql-connector-python==8.3.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
